Moved the kafka API stuff to a seperate KafkaClient class.
